apps/speech/requirements.txt
speechrecognition==3.10.0
google-generativeai==0.3.2
PyAudio==0.2.13